refactor(v3): remove v1/v2 legacy code, migrate utils to v3
- Delete all v2/ (API, Session, Scheduler, Decrypter, PostProcessors, etc.)
- Delete v1 root-level files (Main1, Event, Room, SegmentStreamDownloader, etc.)
- Delete old utils/ directory
- Migrate ClientManager, JsonPath (PathSingle/asString/etc.), withRetry
  to v3/utils/ with updated package names
- Fix all imports from github.rikacelery.utils to github.rikacelery.v3.utils
