--- src/main/kotlin/github/rikacelery/v2/exceptions/DeletedException.kt
@@ -1,3 +1,0 @@
-package github.rikacelery.v2.exceptions
-
-class DeletedException(val name: String):Throwable("model($name) already deleted")
