--- src/main/kotlin/github/rikacelery/v2/exceptions/RenameException.kt
@@ -1,3 +1,0 @@
-package github.rikacelery.v2.exceptions
-
-class RenameException(val newName: String):Throwable("Rename to $newName")
