--- src/main/kotlin/github/rikacelery/v2/postprocessors/ProcessorCtx.kt
@@ -1,13 +1,0 @@
-package github.rikacelery.v2.postprocessors
-
-import github.rikacelery.Room
-import java.util.*
-
-data class ProcessorCtx(
-    val room: Room,
-    val startTime: Date,
-    val endTime: Date,
-    // milliseconds
-    val duration: Long,
-    val quality: String,
-)
